+++ src/types/query.ts
@@ -204,7 +204,7 @@ export interface PaginatedDreamQueryResult<T extends Dream> {
   results: T[]
 }
 
-export interface ScrollPaginatedDreamQueryOptions {
+export interface CursorPaginatedDreamQueryOptions {
   /**
    * the number of records you would like to be returned in each page
    */
@@ -216,7 +216,7 @@ export interface ScrollPaginatedDreamQueryOptions {
   cursor: string | null | undefined
 }
 
-export interface ScrollPaginatedDreamQueryResult<T extends Dream> {
+export interface CursorPaginatedDreamQueryResult<T extends Dream> {
   /**
    * identifier of where to start the next page; undefined to start from the beginning; null when no more pages
    */